Address review comments from @Anthchirp
Move old Python versions of masking code to test files since they
are only used in tests to compare against new C++ code. These could
probably be removed entirely in due course.
